requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.24.4
scipy==1.10.1
pydantic==2.7.4
pydantic-core==2.18.4
sentence-transformers==2.6.1
scikit-learn==1.3.2
networkx==3.1
pyarrow==14.0.2
typer==0.12.3
pyyaml==6.0.1
pytest==7.4.4
